requirements.txt
streamlit>=1.30.0
azure-ai-documentintelligence>=1.0.0
openai>=1.12.0
pydantic>=2.5.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
